+++ trunk/simba/java/junit/sim/scheduling/DispatcherTest.java
@@ -13,13 +13,13 @@ import sim.model.Job;
 
 public class DispatcherTest
 {
-	
+
 	@Test
 	public void testDispatch()
 	{
 		int currentTime = 1;
 		int jobLength = 1;
-		Job job = Job.create((long) jobLength).priority(0).submitTime(0).cores(0).memory(0).build();
+		Job job = Job.create(jobLength).priority(0).submitTime(0).cores(0).memory(0).build();
 		Host host = mock(Host.class);
 		EventQueue eventQueue = new EventQueue(new Clock());
 		new Dispatcher(eventQueue).dipatch(job, host, currentTime);
@@ -29,5 +29,5 @@ public class DispatcherTest
 		Event finish = eventQueue.peek();
 		assertEquals(jobLength + currentTime, finish.time());
 	}
-	
+
 }
