+++ trunk/simba/java/junit/sim/scheduling/matchers/BestFitTest.java
@@ -1,8 +1,7 @@
 package sim.scheduling.matchers;
 
-import static com.google.common.collect.Lists.newArrayList;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static com.google.common.collect.Lists.*;
+import static org.junit.Assert.*;
 
 import java.util.List;
 
@@ -17,7 +16,7 @@ public class BestFitTest
 	public void testNoAccpetingHostNotFit()
 	{
 		List<Host> hosts = newArrayList(Host.create().build());
-		assertNull(new BestFit().match(Job.create((long) 1).cores(1).build(), hosts));
+		assertNull(new BestFit().match(Job.create(1).cores(1).build(), hosts));
 	}
 
 	@Test
@@ -25,7 +24,7 @@ public class BestFitTest
 	{
 		Host host = Host.create().cores(1).build();
 		List<Host> hosts = newArrayList(host);
-		assertEquals(host, new BestFit().match(Job.create((long) 1).cores(1).build(), hosts));
+		assertEquals(host, new BestFit().match(Job.create(1).cores(1).build(), hosts));
 	}
 
 	@Test
@@ -35,7 +34,7 @@ public class BestFitTest
 		Host host2 = Host.create().cores(2).memory(15).build();
 		Host host3 = Host.create().cores(2).memory(8).build();
 		List<Host> hosts = newArrayList(host1, host2, host3);
-		assertEquals(host3, new BestFit().match(Job.create((long) 1).cores(1).memory(4).build(), hosts));
+		assertEquals(host3, new BestFit().match(Job.create(1).cores(1).memory(4).build(), hosts));
 	}
 
 }
